dementia-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
openai==1.3.7
azure-cognitiveservices-speech==1.34.0
tiktoken==0.5.2
requests==2.31.0
pydantic==2.5.0

# 데이터베이스 관련
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1

# 파일 처리
Pillow==10.1.0
python-magic==0.4.27

# 유틸리티
uuid==1.30